refactor(eq): extract the passive EQ pipeline into services/eq
