+++ extension/src/extension/services/eq/lintDenylist.ts
